requirements.txt
python-arq
TgCrypto
aiohttp
aiofiles
future
psutil
pillow
pyrogram
speedtest-cli
uvloop
youtube_dl
search_engine_parser
cryptography
googletrans==4.0.0-rc1
motor
dnspython
python-dotenv